+++ CHANGES.md
@@ -2,6 +2,10 @@
 
 ### 1.95 - 2022-07-01
 
+##### Additions :tada:
+
+- Added support for rendering individual models in 2D / CV using `ModelExperimental`. [#10419](https://github.com/CesiumGS/cesium/pull/10419)
+
 ##### Fixes :wrench:
 
 - Fixed `Cesium3DTileColorBlendMode.REPLACE` for certain tilesets. [#10424](https://github.com/CesiumGS/cesium/pull/10424)